move around components and split off hooks
