+++ app/javascript/mastodon/components/account_header/fields.tsx
@@ -7,11 +7,11 @@ import classNames from 'classnames';
 
 import IconVerified from '@/images/icons/icon_verified.svg?react';
 import { openModal } from '@/mastodon/actions/modal';
-import type { AccountField } from '@/mastodon/features/account_timeline/common';
 import { useFieldHtml } from '@/mastodon/features/account_timeline/hooks/useFieldHtml';
 import { cleanExtraEmojis } from '@/mastodon/features/emoji/normalize';
 import { useAccount } from '@/mastodon/hooks/useAccount';
 import { useResizeObserver } from '@/mastodon/hooks/useObserver';
+import type { AccountFieldShape } from '@/mastodon/models/account';
 import { useAppDispatch } from '@/mastodon/store';
 import MoreIcon from '@/material-icons/400-24px/more_horiz.svg?react';
 
@@ -36,6 +36,12 @@ const dateFormatOptions: Intl.DateTimeFormatOptions = {
   hour: '2-digit',
   minute: '2-digit',
 };
+
+export interface AccountField extends AccountFieldShape {
+  nameHasEmojis: boolean;
+  value_plain: string;
+  valueHasEmojis: boolean;
+}
 
 export const AccountHeaderFields: FC<{ accountId: string }> = ({
   accountId,
