--- app/javascript/mastodon/features/account_timeline/common.ts
@@ -1,7 +1,0 @@
-import type { AccountFieldShape } from '@/mastodon/models/account';
-
-export interface AccountField extends AccountFieldShape {
-  nameHasEmojis: boolean;
-  value_plain: string;
-  valueHasEmojis: boolean;
-}
