+++ app/javascript/mastodon/features/account_timeline/components/filters.tsx
@@ -12,8 +12,8 @@ import { Toggle } from '@/mastodon/components/form_fields';
 import { Icon } from '@/mastodon/components/icon';
 import KeyboardArrowDownIcon from '@/material-icons/400-24px/keyboard_arrow_down.svg?react';
 
-import { useAccountContext } from './context';
-import classes from './styles.module.scss';
+import { useAccountContext } from '../hooks/useAccountContext';
+import classes from '../styles.module.scss';
 
 export const AccountFilters: FC = () => {
   const { acct } = useParams<{ acct: string }>();
