+++ app/javascript/mastodon/features/account_timeline/modals/field_modal.tsx
@@ -2,6 +2,7 @@ import type { FC } from 'react';
 
 import { FormattedMessage } from 'react-intl';
 
+import type { AccountField } from '@/mastodon/components/account_header/fields';
 import { Button } from '@/mastodon/components/button';
 import { EmojiHTML } from '@/mastodon/components/emoji/html';
 import {
@@ -10,7 +11,6 @@ import {
   ModalShellBody,
 } from '@/mastodon/components/modal_shell';
 
-import type { AccountField } from '../common';
 import { useFieldHtml } from '../hooks/useFieldHtml';
 
 import classes from './styles.module.scss';
